Use CSI_SPEC v0.2.0 as current default
